nitial work to support multiple servers
This PR isolates server data from other api info.  At the moment this still acts as a single plex server, but is stored in its own table to allow for future updates where multiple servers will be allows.  This does effectively kill the option to use .env for a single server.  will circle back on the env issue.